Add Product and Home page a11y tests
