+++ src/views/collection/Seo/index.tsx
@@ -1,8 +1,9 @@
 import { graphql } from 'gatsby'
 import { BreadcrumbJsonLd, GatsbySeo } from 'gatsby-plugin-next-seo'
-import React, { useEffect } from 'react'
+import React from 'react'
 import type { CollectionSeoFragment_StoreCollectionFragment } from '@generated/CollectionSeoFragment_storeCollection.graphql'
 import type { CollectionSeoFragment_SiteFragment } from '@generated/CollectionSeoFragment_site.graphql'
+import useAddHead from 'src/utils/hooks/useAddHeadin'
 
 import { useMetadata } from './hooks/useMetadata'
 
@@ -25,23 +26,7 @@ function Seo({
     canonical: `/${slug}/`,
   })
 
-  useEffect(() => {
-    // Add h1 with the title page to fix a11y page-has-heading-one error.
-    const h1 = document.createElement('h1')
-
-    h1.textContent = collectionSeo.title
-    // remove visually, but remain "visible" for screen readers.
-    h1.style.position = 'absolute'
-    h1.style.top = '-100px'
-
-    const main = document.querySelector('main')
-
-    main?.insertBefore(h1, main.childNodes[0])
-
-    return () => {
-      h1.remove()
-    }
-  }, [collectionSeo.title])
+  useAddHead({ title: collectionSeo.title })
 
   return (
     <>
