+++ src/views/home/Seo/index.tsx
@@ -1,5 +1,6 @@
 import { GatsbySeo, JsonLd } from 'gatsby-plugin-next-seo'
 import React from 'react'
+import useAddHead from 'src/utils/hooks/useAddHeadin'
 
 import type { Props as ViewProps } from '../index'
 import { useMetadata } from './hooks/useMetadata'
@@ -10,6 +11,8 @@ export type Props = ViewProps
 function Seo(props: Props) {
   const metadata = useMetadata(props)
   const siteLinksSearchBox = useSiteLinksSearchBoxJsonLd(props)
+
+  useAddHead({ title: metadata.title ?? '' })
 
   return (
     <>
